backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
pydantic>=2.9.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
python-multipart>=0.0.12
aiosqlite>=0.20.0
fastmcp>=0.3.0
websockets>=13.0
python-dotenv>=1.0.1
python-socketio>=5.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=4.5.0
slowapi>=0.1.9
alembic>=1.13.0

# MCP dependencies
openapi-core>=0.19.5
openapi-pydantic>=0.5.1
authlib>=1.5.2
cyclopts>=3.0.0
pyperclip>=1.9.0

# Development dependencies (optional)
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
black>=24.10.0
mypy>=1.13.0
ruff>=0.7.0
